+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - Important - changing import declaration in `package.json` in order to fix tree-shaking issue - [Issue #647](https://github.com/selfrefactor/rambda/issues/647)
 
+- Synchronize with `@types/ramda` - `R.prop`, `R.path`, `R.pickAll`
+
 - Add `R.modify`
 
 - Allow multiple inputs in Typescript versions of `R.anyPass` and `R.allPass` - [Issue #642](https://github.com/selfrefactor/rambda/issues/604)